:truck: move send message handler to separate library
