+++ angular.json
@@ -635,6 +635,31 @@
       },
       "tags": []
     },
+    "functions-bot-engine-send-message": {
+      "root": "libs/functions/bot-engine/send-message",
+      "sourceRoot": "libs/functions/bot-engine/send-message/src",
+      "projectType": "library",
+      "targets": {
+        "lint": {
+          "executor": "@nrwl/linter:eslint",
+          "outputs": ["{options.outputFile}"],
+          "options": {
+            "lintFilePatterns": [
+              "libs/functions/bot-engine/send-message/**/*.ts"
+            ]
+          }
+        },
+        "test": {
+          "executor": "@nrwl/jest:jest",
+          "outputs": ["coverage/libs/functions/bot-engine/send-message"],
+          "options": {
+            "jestConfig": "libs/functions/bot-engine/send-message/jest.config.ts",
+            "passWithNoTests": true
+          }
+        }
+      },
+      "tags": []
+    },
     "functions-bot-engine-whatsapp": {
       "root": "libs/functions/bot-engine/whatsapp",
       "sourceRoot": "libs/functions/bot-engine/whatsapp/src",
