+++ libs/functions/bot-engine/main/src/index.ts
@@ -2,10 +2,8 @@ export * from './lib/engine-bot-manager.handler';
 export * from './lib/io/incoming-message.parser';
 export * from './lib/io/outgoing-message.parser';
 export * from './lib/model/active-channel.service';
-export * from './lib/model/send-message/send-message-main.model';
 export * from './lib/services/data-services/channel-info.service';
 export * from './lib/services/data-services/messages.service';
-export * from './lib/handlers/send-outgoing-message.handler';
 export * from './lib/io/incoming-message-parsers/text-message-parser.class';
 export * from './lib/io/incoming-message-parsers/audio-message-parser.class';
 export * from './lib/io/incoming-message-parsers/video-message-parser.class';
