+++ libs/functions/bot-engine/messenger/src/index.ts
@@ -1,2 +1,2 @@
 export * from './lib/messenger-receive-message.handler';
-// export * from './lib/models/messenger-active-channel.model';
+export * from './lib/models/messenger-active-channel.model';
