+++ libs/functions/bot-engine/send-message/README.md
@@ -1,0 +1,11 @@
+# functions-bot-engine-send-message
+
+This library was generated with [Nx](https://nx.dev).
+
+## Running unit tests
+
+Run `nx test functions-bot-engine-send-message` to execute the unit tests via [Jest](https://jestjs.io).
+
+## Running lint
+
+Run `nx lint functions-bot-engine-send-message` to execute the lint via [ESLint](https://eslint.org/).
