+++ libs/functions/bot-engine/send-message/src/index.ts
@@ -1,0 +1,1 @@
+export * from './lib/send-outgoing-message.handler';
