+++ libs/functions/bot-engine/send-message/src/lib/factories/active-channel/active-channel.factory.ts
@@ -1,9 +1,9 @@
 import { HandlerTools } from "@iote/cqrs";
 
 import { WhatsappActiveChannel } from "@app/functions/bot-engine/whatsapp";
+import { MessengerActiveChannel } from "@app/functions/bot-engine/messenger";
 
 import { CommunicationChannel, PlatformType } from "@app/model/convs-mgr/conversations/admin/system";
-// import { MessengerActiveChannel } from "@app/functions/bot-engine/messenger";
 
 /**
  * @Description When an end user sends a message to the chatbot from a thirdparty application, we 
@@ -21,8 +21,8 @@ export class ActiveChannelFactory
     switch (communicationChannel.type) {
       case PlatformType.WhatsApp:
         return new WhatsappActiveChannel(tools, communicationChannel);
-      // case PlatformType.Messenger:
-      //   return new MessengerActiveChannel(tools, communicationChannel);
+      case PlatformType.Messenger:
+        return new MessengerActiveChannel(tools, communicationChannel);
       default:
         return new WhatsappActiveChannel(tools, communicationChannel);
     }
