+++ tsconfig.base.json
@@ -177,6 +177,9 @@
       "@app/functions/bot-engine/messenger": [
         "libs/functions/bot-engine/messenger/src/index.ts"
       ],
+      "@app/functions/bot-engine/send-message": [
+        "libs/functions/bot-engine/send-message/src/index.ts"
+      ],
       "@app/functions/bot-engine/whatsapp": [
         "libs/functions/bot-engine/whatsapp/src/index.ts"
       ],
